refactor: add components to main component
