+++ src/pages/main/Main.tsx
@@ -3,15 +3,15 @@ import {ArticleInterface, EventInterface, TagsInterface} from "@/types";
 import { useOnlineCheck } from "@/shared/hooks/useOnlineCheck";
 import { BadInternetConnection } from "@/components/BadInternetConnection/BadInternetConnection";
 import { Navbar } from "@/components/Navbar/Navbar";
-import {RoutineSlider} from "@/components/RoutineSlider/RoutineSlider.tsx";
-import {MoodButton} from "@/components/MoodButton/MoodButton.tsx";
+import {RoutineSlider} from "@/pages/main/components/RoutineSlider/RoutineSlider.tsx";
+import {MoodButton} from "@/pages/main/components/MoodButton/MoodButton.tsx";
 import {BurnoutTestBanner} from "@/components/BurnoutTestBanner/BurnoutTestBanner.tsx";
 import {MoodGraph} from "@/components/MoodGraph/MoodGraph.tsx";
 import {BurnoutLevel} from "@/components/BurnoutLevel/BurnoutLevel.tsx";
 import {PsychologistInfo} from "@/components/PsychologistInfo/PsychologistInfo.tsx";
 import {Articles} from "@/components/Articles/Articles.tsx";
 import {EventsMain} from "@/components/EventsMain/EventsMain.tsx";
-import {PieChart} from "@/components/PieChart/PieChart.tsx";
+import {PieChart} from "@/pages/main/components/PieChart/PieChart.tsx";
 import {useRequest} from "@/shared/hooks/useRequest.tsx";
 import {getActivityTypes} from "@/shared/api/Api.ts";
 import {ReactElement, useState} from "react";
