+++ src/pages/main/components/MoodButton/MoodButton.tsx
