+++ src/pages/main/components/PieChart/components/PeriodPicker.module.scss
@@ -1,4 +1,4 @@
-@import '@/shared/styles/mixins.scss';
+@import '@/shared/styles/mixins';
 
 .period-picker{
   width: 156px;
@@ -22,9 +22,9 @@
     @include background-no-repeat-center-cover;
     border: none;
     background-color: var(--color-background);
-    background-image: url('../assets/chevron_down_24.svg');
+    background-image: url('chevron_down_24.svg');
     &-opened {
-      background-image: url('../assets/chevron_up_24.svg');
+      background-image: url('chevron_up_24.svg');
     }
   }
 }
