+++ src/pages/main/components/RoutineSlider/RoutineSlider.helpers.ts
