+++ src/pages/main/components/RoutineSlider/components/ButtonsList/ButtonsList.module.scss
@@ -1,4 +1,4 @@
-@import "@/shared/styles/mixins.scss";
+@import "@/shared/styles/mixins";
 
 .buttons-list {
   margin: 24px 0 0 38px;
