+++ src/pages/main/components/RoutineSlider/components/ButtonsList/ButtonsList.tsx
