+++ src/pages/main/components/RoutineSlider/components/RoutineSection/RoutineSection.module.scss
