+++ src/pages/main/components/RoutineSlider/routineslider.module.scss
@@ -1,4 +1,4 @@
-@import "@/shared/styles/mixins.scss";
+@import "@/shared/styles/mixins";
 
 .container {
   background-color: var(--color-background-white);
